% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCuts.R
\name{checkCuts}
\alias{checkCuts}
\title{checkCuts}
\usage{
checkCuts(cutSites, cutIDs, genome, fasta = FALSE, seq)
}
\arguments{
\item{cutSites}{A GRanges object containing the locations of the cut sites to be checked for sequence match.}

\item{cutIDs}{A character vector containing unique IDs for each cut site.}

\item{genome}{The path to a fasta file or a BSgenome object to check for genomic sequences.}

\item{fasta}{TRUE if a fasta file has been supplied. Default = FALSE}

\item{seq}{The desired recognition sequence that the enzyme should have cut.}
}
\value{
cutIDs that match the desired recognition sequence (seq)
}
\description{
Determines the sequence around a cut site using a fasta file or BSgenome
}
\examples{
library(GenomicRanges)
# Load the positions of possible MspI cut sites
data(datCounts)
cutSites <- GRanges(rownames(datCounts))
# Adjust positions to overlap the recognition sequence of MspI
start(cutSites) <- start(cutSites) - 1
end(cutSites) <- end(cutSites) + 2
chr20 <- system.file("extdata", "chr20.fa.gz", package="msgbsR")
correctCuts <- checkCuts(cutSites = cutSites, cutIDs = row.names(datCounts),
                         genome = chr20, fasta = TRUE, seq = "CCGG")
}
\author{
Benjamin Mayne
}

