% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closestFeat.R
\name{closestFeat}
\alias{closestFeat}
\title{closestFeat}
\usage{
closestFeat(gff, cutSites)
}
\arguments{
\item{gff}{The path to the location of a gff file.}

\item{cutSites}{A GRanges object containing the location of the cut sites.}
}
\value{
The closest biological feature annotated in a gff3 file.
}
\description{
Determines the closest feature of cut site locations
}
\examples{
data(cuts)
mygff <- system.file("extdata", "chr20.gff3", package = 'msgbsR')
features <- closestFeat(gff = mygff, cutSites = cuts)
}
\author{
Benjamin Mayne
}

