% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgbsR.R
\docType{data}
\name{cuts}
\alias{cuts}
\title{A GRanges object of differentially methylated MspI cut sites from a MS-GBS experiment.}
\format{A GRanges object of length 363.}
\usage{
data(cuts)
}
\value{
A GRanges object of length 363.
}
\description{
The GRanges object was created from a list of differentially methylated cut sites from a MS-GBS experiment between two groups of rats that were fed different diets.
}
\details{
\itemize{
	\item Positions of MspI cut sites differentially methylated on chromosome 20 in Rats.
}
}
\keyword{datasets}

