% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgbsR.R
\docType{data}
\name{datCounts}
\alias{datCounts}
\title{Table of read counts}
\format{A 16977x6 matrix}
\usage{
data(datCounts)
}
\value{
A 16977x6 matrix
}
\description{
A data set containing the following data:
}
\details{
\itemize{
	\item datCounts a matrix with 16977 rows (Potential MspI cut sites) and six columns (samples)
}
}
\keyword{datasets}

