% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgbsR.R
\docType{data}
\name{datCounts_filtered}
\alias{datCounts_filtered}
\title{Table of filtered read counts from a MSGBS experiment}
\format{A 7178x6 matrix}
\usage{
data(datCounts_filtered)
}
\value{
A 7178x6 matrix
}
\description{
A data set containing the following data:
}
\details{
\itemize{
	\item datCounts_filtered a matrix with 7178 rows (Correct MspI cut sites) and six columns (samples)
}
}
\keyword{datasets}

