% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotChr.R
\name{plotChr}
\alias{plotChr}
\title{plotChr}
\usage{
plotChr(cutSites, genome)
}
\arguments{
\item{cutSites}{A matrix consisting of three columns, where the first column is the chromosome ID, the second and third column being the start and end position of the cut site respectively.}

\item{genome}{A matrix with the lengths of the chromosomes. The first column is the chromosome ID and the second is the length of each chromosome.}
}
\value{
An ideogram showing the locations of the cut sites.
}
\description{
Plot a karogram representing the cut site locations
}
\examples{
chr20 <- matrix(c('chr20', '56205956'), nrow = 1, ncol = 2)
data <- system.file("extdata", "exampleSites.Rdata", package = 'msgbsR')
load(data)
plotChr(cutSites = z[,c(1,3,3)], genome = chr20)
}
\author{
Benjamin Mayne
}

