% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCircos.R
\name{plotCircos}
\alias{plotCircos}
\title{plotCircos}
\usage{
plotCircos(cutSites, genome, cutSite.colour, genome.colour)
}
\arguments{
\item{cutSites}{A matrix consisting of three columns, where the first column is the chromosome ID, the second and third column being the start and end position of the cut site respectively.}

\item{genome}{A matrix with the lengths of the chromosomes. The first column is the chromosome ID and the second is the length of each chromosome.}

\item{cutSite.colour}{The colour of the cut sites.}

\item{genome.colour}{The colour of the chromosomes}
}
\description{
Plot a circos representing the cut site locations
}
\examples{
\dontrun{
my_cuts <- matrix(c('chr1', 'chr2', 'chr3', 'chr4',
'500', '1000', '750', '400',
'500', '1000', '750', '400'), nrow=4, ncol=3)
my_genome <- matrix(c('chr1', 'chr2', 'chr3', 'chr4',
                   '1000', '2000','1500', '800'), nrow=4, ncol=3)
plotCircos(cutSites = my_cuts, genome = my_genome, cutSite.colour = 'red', genome.colour = 'green')

}
}
\author{
Benjamin Mayne
}

