% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCounts.R
\name{plotCounts}
\alias{plotCounts}
\title{plotCounts}
\usage{
plotCounts(countMatrix, condition1, condition2 = NULL)
}
\arguments{
\item{countMatrix}{A data frame containing read counts per sample.}

\item{condition1}{A character vector containing a certain condition of the samples.}

\item{condition2}{A character vector containing another certain condition of the samples. Default is NULL.}
}
\value{
Produces a plot showing the total number reads vs total number of cut site per sample.
}
\description{
Plot the total number of reads vs total number of cut sites per sample
}
\examples{
# Generate a random matrix
set.seed(1)
x <- matrix(sample(0:100, size = 10000*10, replace = TRUE), nrow = 10000, ncol = 10)
y <- c(rep('A', 5), rep('B', 5))
z <- c(rep('C', 3), rep('D', 2), rep('E', 2), rep('F', 3))

plotCounts(x,y,z)
}
\author{
Benjamin Mayne
}

