% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgbsR.R
\docType{data}
\name{ratChr}
\alias{ratChr}
\title{A integer with the length of chr20 in Rat}
\format{A integer of length 1}
\usage{
data(ratChr)
}
\value{
A integer of length 1
}
\description{
A data set containing the following data:
}
\details{
\itemize{
	\item ratChr an integer of the value 56205956
}
}
\keyword{datasets}

