% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgbsR.R
\docType{data}
\name{ratdata}
\alias{ratdata}
\title{Read counts of potential MspI cut sites from a MS-GBS experiment of prostates from rats}
\format{RangedSummarizedExperiment}
\usage{
data(ratdata)
}
\value{
RangedSummarizedExperiment
}
\description{
A RangedSummarizedExperiment containing read counts generated from a MS-GBS experiment using the restriction enzyme MspI, focusing on chromosome 20 of Rat.
}
\details{
\itemize{
	\item ratdata A RangedSummarizedExperiment with 16047 potential MspI cut sites on chromosome 20 in Rat and six samples (3 Control and 3 Experimental).
}

This dataset contains six prostate samples from rats: 3 control and 3 experimental high fat diet.
}
\keyword{datasets}

