% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgbsR.R
\docType{data}
\name{ratdata2}
\alias{ratdata2}
\title{Read counts of correct MspI cut sites from a MS-GBS experiment of prostates from rats}
\format{RangedSummarizedExperiment}
\usage{
data(ratdata2)
}
\value{
RangedSummarizedExperiment
}
\description{
A RangedSummarizedExperiment containing read counts generated from a MS-GBS experiment using the restriction enzyme MspI, focusing on chromosome 20 of Rat. The sites have been checked for the correct recognition site.
}
\details{
\itemize{
	\item ratdata2 A RangedSummarizedExperiment containing data for 13983 MspI cut sites on chromosome 20 in Rat and six samples (3 Control and 3 Experimental).
}

This dataset contains six prostate samples from rats: 3 control and 3 experimental high fat diet. The data can be used for differential methylation analyses.
}
\keyword{datasets}

