% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCuts.R
\name{checkCuts}
\alias{checkCuts}
\title{checkCuts}
\usage{
checkCuts(cutSites, cutIDs, genome, fasta = FALSE, seq)
}
\arguments{
\item{cutSites}{A matrix where the first column is the chromosome ID, the second column is the starting position and the third column is the ending position of the cut site sequence.}

\item{cutIDs}{A character vector containing unique IDs for each cut site.}

\item{genome}{The path to fasta file or a BSgenome object to check for genomic sequences.}

\item{fasta}{TRUE if a fasta file has been supplied. Default = FALSE}

\item{seq}{The desired sequence that the enzyme should have cut.}
}
\value{
cut sites that match the input sequence
}
\description{
Determines the sequence around a cut site using a fasta file
}
\examples{
library(GenomicRanges)
data(datCounts)
cutSites <- GRanges(rownames(datCounts))
cutSites@ranges@start <- as.integer(cutSites@ranges@start -1)
cutSites@ranges@width <- as.integer(cutSites@ranges@width + 3)
chr20 <- system.file("extdata", "chr20.fa.gz", package="msgbsR")
correctCuts <- checkCuts(cutSites = cutSites, cutIDs = row.names(datCounts),
                         genome = chr20, fasta = TRUE, seq = 'CCGG')
}
\author{
Benjamin Mayne
}

