% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closestFeat.R
\name{closestFeat}
\alias{closestFeat}
\title{closestFeat}
\usage{
closestFeat(gff, positions, strand.specific = TRUE)
}
\arguments{
\item{gff}{The path to the location of a gff file.}

\item{positions}{A matirx where the first column is the chromosome ID, the second column is the strand (1 or 2) and the third column is the position.}

\item{strand.specific}{If TRUE the function will return the closest feature on the same strand, otherwise it'll return the closest feature on either strand. Default is TRUE.}
}
\description{
Determines the closest feature of cut site locations
}
\author{
Benjamin Mayne
}

