% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgbsR.R
\docType{data}
\name{cuts}
\alias{cuts}
\title{A GRanges object of MspI cut sites.}
\format{A GRanges object}
\usage{
data(cuts)
}
\value{
A GRanges object
}
\description{
A data set containing the following data:
}
\details{
\itemize{
	\item cuts a GRanges object of MspI cut sites
}
}
\keyword{datasets}

