% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgbsR.R
\docType{data}
\name{datCounts}
\alias{datCounts}
\title{A table of read counts from a MS-GBS experiment}
\format{A 16977x6 integer matrix}
\usage{
data(datCounts)
}
\value{
A 16977x6 integer matrix
}
\description{
An integer matrix containing read counts generated from a MS-GBS experiment using the restriction enzyme MspI, focusing on chromosome 20 of Rat.
}
\details{
\itemize{
	\item datCounts a matrix with 16977 rows (Potential MspI cut sites on chromosome 20) and six columns (samples).
}
}
\keyword{datasets}

