% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgbsR.R
\docType{data}
\name{datCounts_filtered}
\alias{datCounts_filtered}
\title{A table of read counts from a MS-GBS experiment after filtering incorrect cut sites}
\format{A 7178x6 integer matrix}
\usage{
data(datCounts_filtered)
}
\value{
A 7178x6 integer matrix
}
\description{
An integer matrix of read counts with correct locations of MspI cuts of chromosome 20 from Rat.
}
\details{
\itemize{
	\item datCounts_filtered a matrix with 7178 rows (Correct MspI cut sites) and six columns (samples)
}
}
\keyword{datasets}

