% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotChr.R
\name{plotChr}
\alias{plotChr}
\title{plotChr}
\usage{
plotChr(cutSites, genome)
}
\arguments{
\item{cutSites}{A matrix consisting of three columns, where the first column is the chromosome ID, the second and third column being the start and end position of the cut site respectively.}

\item{genome}{A matrix with the lengths of the chromosomes. The first column is the chromosome ID and the second is the length of each chromosome.}
}
\value{
An ideogram showing the locations of the cut sites.
}
\description{
Plot a karogram representing the cut site locations
}
\examples{
my_cuts <- matrix(c('chr1', 'chr2', 'chr3', 'chr4',
'500', '1000', '750', '400',
'500', '1000', '750', '400'), nrow=4, ncol=3)
my_genome <- matrix(c('chr1', 'chr2', 'chr3', 'chr4',
                   '1000', '2000','1500', '800'), nrow=4, ncol=3)
plotChr(cutSites = my_cuts, genome = my_genome)

}
\author{
Benjamin Mayne
}

