% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCircos.R
\name{plotCircos}
\alias{plotCircos}
\title{plotCircos}
\usage{
plotCircos(cutSites, genome, cutSite.colour, genome.colour)
}
\arguments{
\item{cutSites}{A matrix consisting of three columns, where the first column is the chromosome ID, the second and third column being the start and end position of the cut site respectively.}

\item{genome}{An integer with the lengths of the chromosomes.}

\item{cutSite.colour}{The colour of the cut sites.}

\item{genome.colour}{The colour of the chromosomes}
}
\value{
A circos plot showing the locations of the cut sites.
}
\description{
Plot a circos representing the cut site locations
}
\examples{
data(cuts)
data(ratChr)
plotCircos(cutSites = cuts, genome = ratChr,
           cutSite.colour = 'red', genome.colour = 'blue')
}
\author{
Benjamin Mayne
}

