% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCircos.R
\name{plotCircos}
\alias{plotCircos}
\title{plotCircos}
\usage{
plotCircos(cutSites, seqlengths, cutSite.colour, seqlengths.colour)
}
\arguments{
\item{cutSites}{A GRanges object containing the locations of the cut sites.}

\item{seqlengths}{An integer with the lengths of the chromosomes.}

\item{cutSite.colour}{The colour of the cut sites.}

\item{seqlengths.colour}{The colour of the chromosomes}
}
\value{
A circos plot showing the locations of the cut sites.
}
\description{
Plot a circos representing the cut site locations
}
\examples{
data(cuts)
data(ratChr)
plotCircos(cutSites = cuts, seqlengths = ratChr,
           cutSite.colour = "red", seqlengths.colour = "blue")
}
\author{
Benjamin Mayne
}

