% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCounts.R
\name{plotCounts}
\alias{plotCounts}
\title{plotCounts}
\usage{
plotCounts(se, cateogory)
}
\arguments{
\item{se}{A RangedSummarizedExperiment containing meta data of the samples.}

\item{cateogory}{The heading name in the sample data to distinguish groups.}
}
\value{
Produces a plot showing the total number reads vs total number of cut sites per sample.
}
\description{
Plots the total number of reads vs total number of cut sites per sample
}
\examples{
data(ratdata2)
plotCounts(se = ratdata2, cateogory = "Group")
}
\author{
Benjamin Mayne
}

