% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCounts.R
\name{plotCounts}
\alias{plotCounts}
\title{plotCounts}
\usage{
plotCounts(countMatrix, condition1, condition2 = NULL)
}
\arguments{
\item{countMatrix}{A matrix containing read counts per sample.}

\item{condition1}{A character vector containing a certain condition of the samples.}

\item{condition2}{A character vector containing another certain condition of the samples. Default is NULL.}
}
\value{
Produces a plot showing the total number reads vs total number of cut site per sample.
}
\description{
Plot the total number of reads vs total number of cut sites per sample
}
\examples{
data(datCounts_filtered)
y <- data.frame(c(rep("Control", 3), rep("Fat Diet", 3)))
colnames(y) <- "Group"
plotCounts(countMatrix = datCounts, condition1 = y$Group)
}
\author{
Benjamin Mayne
}

