% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgbsR.R
\docType{data}
\name{ratChr}
\alias{ratChr}
\title{The length of chromsome 20 in Rats (rn6 UCSC)}
\format{An integer of length 1}
\usage{
data(ratChr)
}
\value{
An integer of length 1
}
\description{
An integer with the length of base pairs of chr20 in Rat (rn6 UCSC).
}
\details{
\itemize{
	\item ratChr an integer of the value 56205956
}
}
\keyword{datasets}

