% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawCounts.R
\name{rawCounts}
\alias{rawCounts}
\title{rawCounts}
\usage{
rawCounts(bamFilepath, threads = 1)
}
\arguments{
\item{bamFilepath}{The path to the location of bam file(s).}

\item{threads}{The total number of usable threads to be used. Default is 1.}
}
\value{
Produces a data frame where the rows correspond to cut sites and columns are the samples. Cut sites are given a unique ID (chromosome:strand:position).
}
\description{
Imports the raw read counts from sorted and indexed bam file(s)
}
\examples{
my_path <- system.file("extdata", "Control_1.bam", package = 'msgbsR')
my_path <- gsub('Control_1.bam', '', my_path)
datCounts <- rawCounts(bamFilepath = my_path, threads = 1)
}
\author{
Benjamin Mayne, Sam Buckberry
}

